// types/hortzettel.ts
export interface Child {
  id: string;
  firstName: string;
  lastName: string;
  class?: string;
  birthDate?: string;
  allergies?: string;
  medicalNotes?: string;
  authorizedPickup?: string;
}